devops/kubernates/requirements.txt
click==8.1.7
h11==0.14.0
Mako==1.2.4
MarkupSafe==2.1.1
python-dotenv==1.0.0
typing_extensions==4.8.0
ujson==5.6.0
